+++ Arius/Arius.Core/Repositories/Repository.Binaries.cs
@@ -1,72 +1,34 @@
-﻿//using Arius.Core.Extensions;
-//using Arius.Core.Models;
-//using Azure;
-//using Azure.Storage.Blobs;
-//using Azure.Storage.Blobs.Models;
-//using Azure.Storage.Blobs.Specialized;
-//using Microsoft.EntityFrameworkCore;
-//using Microsoft.Extensions.Logging;
-//using System;
-//using System.Collections.Generic;
-//using System.IO.Compression;
-//using System.Linq;
-//using System.Net;
-//using System.Text.Json;
-//using System.Threading.Tasks;
+﻿using Arius.Core.Models;
+using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
 
-//namespace Arius.Core.Repositories;
+namespace Arius.Core.Repositories;
 
-//internal partial class Repository
-//{
-//    public BinaryRepository Binaries { get; init; }
+internal partial class Repository
+{
+    /// <summary>
+    /// Get the count of (distinct) BinaryHashes
+    /// </summary>
+    /// <returns></returns>
+    public async Task<int> CountBinariesAsync()
+    {
+        await using var db = GetAriusDbContext();
+        return await db.BinaryProperties.CountAsync();
+        //return await db.PointerFileEntries
+        //    .Select(pfe => pfe.BinaryHash)
+        //    .Distinct()
+        //    .CountAsync();
+    }
 
-//    internal class BinaryRepository
-//    {
-//        private readonly ILogger<Repository>                                   logger;
-//        private readonly Repository                                            repo;
-//        private readonly BlobContainerClient                                   container;
-        
+    public async Task<bool> BinaryExistsAsync(BinaryHash bh)
+    {
+        await using var db = GetAriusDbContext();
+        return await db.BinaryProperties.AnyAsync(bp => bp.Hash == bh);
+    }
 
-//        internal BinaryRepository(Repository parent, BlobContainerClient container)
-//        {
-//            this.logger = parent.logger;
-//            this.repo = parent;
-//            this.container = container;
-//        }
-
-
-
-//        // --- BINARY PROPERTIES ------------------------------------------------
-
-        
-
-        
-
-        
-
-        
-
-        
-
-//        ///// <summary>
-//        ///// Get all the (distinct) BinaryHashes
-//        ///// </summary>
-//        ///// <returns></returns>
-//        //public async Task<BinaryHash[]> GetAllBinaryHashesAsync()
-//        //{
-//        //    await using var db = repo.GetAriusDbContext();
-//        //    return await db.BinaryProperties
-//        //        .Select(bp => bp.Hash)
-//        //        .ToArrayAsync();
-//        //    //return await db.PointerFileEntries
-//        //    //    .Select(pfe => pfe.BinaryHash)
-//        //    //    .Distinct()
-//        //    //    .ToArrayAsync();
-//        //}
-
-//        // --- CHUNKLIST ------------------------------------------------
-
-        
-        
-//    }
-//}
+    public async Task<long> TotalBinaryIncrementalLengthAsync()
+    {
+        await using var db = GetAriusDbContext();
+        return await db.BinaryProperties.SumAsync(bp => bp.IncrementalLength);
+    }
+}
